feat(history): store group history
move history saving and loading to GenericChatForm from ChatForm

Fix #5452
